refactor main.rs into lib.rs and main.rs; add python wrapper for read_config_files() and configure_db()
